@@ -1,14 +1,15 @@
 source 'http://rubygems.org'
+source 'https://fe25782d:2b14c6ee@www.mikeperham.com/rubygems/'
 
-gem 'rails', '~> 4.0.0'
+gem 'rails', '~> 4.1.0'
 
 # Gems used only for assets and not required
 # in production environments by default.
-gem 'sass-rails',   '~> 4.0.1'
+gem 'sass-rails',   '~> 4.0.3'
 gem 'coffee-rails', '~> 4.0.1'
 
 gem 'execjs' #, '~> 1.4.0'
-# gem 'therubyracer'
+# gem 'therubyracer', platforms: :ruby
 # gem 'therubyrhino', '~> 2.0.2'
 
 gem 'uglifier', '~> 2.4.0'
@@ -70,8 +71,7 @@ gem 'retryable-rb', '~> 1.1.0'
 gem 'rollout', '~> 2.0.0'
 gem 'ruby-rtf'
 gem 'secure_headers'
-gem 'sidekiq', '~> 2.17.0'
-gem 'sidekiq-failures', git: 'http://github.com/mhfs/sidekiq-failures.git'
+gem 'sidekiq-pro'
 gem 'sidekiq-unique-jobs'
 gem 'siebel_donations', '~> 1.0.5'
 gem 'sinatra', require: nil
@@ -88,6 +88,7 @@ gem 'global_phone_dbgen'
 #gem 'font_assets'
 gem 'geocoder'
 gem 'google_timezone'
+gem 'email_validator'
 gem 'peek'
 gem 'peek-pg'
 gem 'peek-git'
@@ -108,6 +109,7 @@ end
 group :development, :test do
   gem 'awesome_print'
   gem 'database_cleaner'
+  gem 'spring-commands-rspec'
   gem 'rspec', '~> 2.14.1' #3.0.0.beta2'
   gem 'rspec-rails' #, '3.0.0.beta2'
   gem 'factory_girl_rails'
@@ -123,14 +125,18 @@ group :development, :test do
   gem 'zonebie'
   gem 'equivalent-xml'
 end
+
 group :test do
+  gem 'mock_redis'
   gem 'webmock', '~> 1.9.0'
-  gem 'spork-rails' #, '~> 3.2.0'
   gem 'rb-fsevent', require: false
-  gem 'guard-spork'
   gem 'growl'
   gem 'capybara'
   gem 'resque_spec'
   gem 'emoji_spec', git: 'https://gist.github.com/6112257.git'
-  gem 'rubocop', '~> 0.26.0'
+  gem 'rubocop'
+end
+
+group :doc do
+  gem 'sdoc', '~> 0.4.0'
 end
