@@ -13,7 +13,8 @@ class Api::V1::ContactsController < Api::V1::BaseController
 
     if params[:since]
       meta = {
-        deleted: Version.where(item_type: 'Contact', event: 'destroy', related_object_type: 'AccountList', related_object_id: current_account_list.id).where('created_at > ?', Time.at(params[:since].to_i)).pluck(:item_id),
+        deleted: Version.where(item_type: 'Contact', event: 'destroy', related_object_type: 'AccountList', related_object_id: current_account_list.id)
+                        .where('created_at > ?', Time.at(params[:since].to_i)).pluck(:item_id),
         inactivated: inactivated
       }
     else
