@@ -2,6 +2,15 @@ class ContactExhibit < DisplayCase::Exhibit
   include DisplayCase::ExhibitsHelper
   include ApplicationHelper
 
+  TABS = {
+    'details' => _('Details'),
+    'tasks' => _('Tasks'),
+    'history' => _('History'),
+    'referrals' => _('Referrals'),
+    'notes' => _('Notes'),
+    'social' => _('Social')
+  }
+
   def self.applicable_to?(object)
     object.class.name == 'Contact'
   end
@@ -24,9 +33,10 @@ class ContactExhibit < DisplayCase::Exhibit
   end
 
   def contact_info
-    people.map {|p|
+    people.order('contact_people.primary::int desc').references(:contact_people).map {|p|
       person_exhibit = exhibit(p, @context)
-      [@context.link_to(person_exhibit, @context.contact_person_path(to_model, p)), [person_exhibit.phone_number, person_exhibit.email].compact.map { |e| exhibit(e, @context) }.join('<br />')].select(&:present?).join(':<br />')
+      phone_and_email_exhibits = [person_exhibit.phone_number, person_exhibit.email].compact.map { |e| exhibit(e, @context) }.join('<br />')
+      [@context.link_to(person_exhibit, @context.contact_person_path(to_model, p)), phone_and_email_exhibits].select(&:present?).join(':<br />')
     }.join('<br />').html_safe
   end
 
