@@ -4,7 +4,7 @@ require File.expand_path('../application', __FILE__)
 Encoding.default_external = Encoding::UTF_8
 Encoding.default_internal = Encoding::UTF_8
 # Initialize the rails application.
-Mpdx::Application.initialize!
+Rails.application.initialize!
 
 if defined?(PhusionPassenger)
   PhusionPassenger.on_event(:starting_worker_process) do |forked|
