@@ -1,3 +1,5 @@
+require 'sidekiq/pro/reliable_push'
+
 rails_root = ENV['RAILS_ROOT'] || Rails.root.to_s
 rails_env = ENV['RAILS_ENV'] || 'development'
 
@@ -9,8 +11,8 @@ Sidekiq.configure_client do |config|
 end
 
 Sidekiq.configure_server do |config|
+  require 'sidekiq/pro/reliable_fetch'
   config.redis = { url: 'redis://' + redis_config[rails_env],
                    namespace: "MPDX:#{rails_env}:resque"}
-  config.failures_default_mode = :exhausted
 end
 
