@@ -5,6 +5,11 @@ describe GoogleCalendarIntegrator do
   let(:integrator) { GoogleCalendarIntegrator.new(google_integration) }
   let(:task) { create(:task, account_list: google_integration.account_list, activity_type: 'Appointment') }
   let(:google_event) { create(:google_event, activity: task, google_integration: google_integration) }
+  let(:missing_event_response) {
+    double(data: { 'error' => { 'errors' => [{ 'domain' => 'global', 'reason' => 'notFound', 'message' => 'Not Found' }],
+                                'code' => 404, 'message' => 'Not Found' } },
+      status: 404)
+  }
 
   context '#sync_tasks' do
     it 'calls #sync_task for each future, uncompleted task that is set to be synced' do
@@ -56,7 +61,7 @@ describe GoogleCalendarIntegrator do
 
     it 'removes the calendar integration if the calendar no longer exists on google' do
       google_integration.stub_chain(:calendar_api, :events, :insert).and_return('')
-      integrator.client.should_receive(:execute).and_return(double(data: { 'error' => { 'errors' => [{ 'domain' => 'global', 'reason' => 'notFound', 'message' => 'Not Found' }], 'code' => 404, 'message' => 'Not Found' } }, status: 404))
+      integrator.client.should_receive(:execute).and_return(missing_event_response)
       integrator.should_receive(:event_attributes).and_return({})
 
       integrator.add_task(task)
@@ -84,7 +89,7 @@ describe GoogleCalendarIntegrator do
     it 'adds the google event if it is missing from google' do
       google_integration.stub(:calendars).and_return(nil)
       google_integration.stub_chain(:calendar_api, :events, :patch).and_return('')
-      integrator.client.should_receive(:execute).and_return(double(data: { 'error' => { 'errors' => [{ 'domain' => 'global', 'reason' => 'notFound', 'message' => 'Not Found' }], 'code' => 404, 'message' => 'Not Found' } }, status: 404))
+      integrator.client.should_receive(:execute).and_return(missing_event_response)
       integrator.should_receive(:event_attributes).and_return({})
       integrator.should_receive(:add_task)
 
