@@ -3,9 +3,14 @@ require 'spec_helper'
 describe Person::RelayAccount do
   before(:each) do
     @org = create(:ccc)
-    @auth_hash = Hashie::Mash.new(uid: 'JOHN.DOE@EXAMPLE.COM', extra: { attributes: [{ firstName: 'John', lastName: 'Doe', username: 'JOHN.DOE@EXAMPLE.COM', email: 'johnnydoe@example.com', designation: '0000000', emplid: '000000000', ssoGuid: 'F167605D-94A4-7121-2A58-8D0F2CA6E024' }] })
+    user_attributes = [{ firstName: 'John', lastName: 'Doe', username: 'JOHN.DOE@EXAMPLE.COM',
+                         email: 'johnnydoe@example.com', designation: '0000000', emplid: '000000000',
+                         ssoGuid: 'F167605D-94A4-7121-2A58-8D0F2CA6E024' }]
+    @auth_hash = Hashie::Mash.new(uid: 'JOHN.DOE@EXAMPLE.COM', extra: { attributes: user_attributes })
+    @wsapi_headers = { 'Accept' => '*/*; q=0.5, application/xml', 'Accept-Encoding' => 'gzip, deflate',
+                       'Authorization' => "Bearer #{APP_CONFIG['itg_auth_key']}", 'User-Agent' => 'Ruby' }
     stub_request(:get, 'https://wsapi.ccci.org/wsapi/rest/profiles?response_timeout=60000&ssoGuid=F167605D-94A4-7121-2A58-8D0F2CA6E024')
-      .with(headers: { 'Accept' => '*/*; q=0.5, application/xml', 'Accept-Encoding' => 'gzip, deflate', 'Authorization' => "Bearer #{APP_CONFIG['itg_auth_key']}", 'User-Agent' => 'Ruby' })
+      .with(headers: @wsapi_headers)
       .to_return(status: 200, body: '[]', headers: {})
   end
   describe 'find or create from auth' do
@@ -31,8 +36,9 @@ describe Person::RelayAccount do
 
     it 'creates an organization account if this user has a profile at cru' do
       stub_request(:get, 'https://wsapi.ccci.org/wsapi/rest/profiles?response_timeout=60000&ssoGuid=F167605D-94A4-7121-2A58-8D0F2CA6E024')
-        .with(headers: { 'Accept' => '*/*; q=0.5, application/xml', 'Accept-Encoding' => 'gzip, deflate', 'Authorization' => "Bearer #{APP_CONFIG['itg_auth_key']}", 'User-Agent' => 'Ruby' })
-        .to_return(status: 200, body: '[{"name":"Staff Account(000555555)","designations":[{"number":"0555555","description":"Jon and Jane Doe(000555555)","staffAccountId":"000555555"}]}]', headers: {})
+        .with(headers: @wsapi_headers)
+        .to_return(status: 200, headers: {},
+                   body: '[{"name":"Staff Account(000555555)","designations":[{"number":"0555555","description":"Jon and Jane Doe(000555555)","staffAccountId":"000555555"}]}]')
       person = create(:user)
       @org.stub(:api).and_return(FakeApi.new)
       expect {
@@ -46,8 +52,8 @@ describe Person::RelayAccount do
     it 'should create a user with a first and last name' do
       expect {
         user = Person::RelayAccount.create_user_from_auth(@auth_hash)
-        user.first_name.should == @auth_hash.extra.attributes.first.firstName
-        user.last_name.should == @auth_hash.extra.attributes.first.lastName
+        user.first_name.should eq @auth_hash.extra.attributes.first.firstName
+        user.last_name.should eq @auth_hash.extra.attributes.first.lastName
       }.to change(User, :count).from(0).to(1)
     end
   end
@@ -56,7 +62,7 @@ describe Person::RelayAccount do
     user = create(:user)
     @org.stub(:api).and_return(FakeApi.new)
     Person::RelayAccount.find_or_create_from_auth(@auth_hash, user)
-    Person::RelayAccount.find_authenticated_user(@auth_hash).should == user
+    Person::RelayAccount.find_authenticated_user(@auth_hash).should eq user
   end
 
   it 'should return name for to_s' do
